# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Moru_task",
        "description": "## Quiz Application\nSimple quiz application with user authentication",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "components": {
        "securitySchemes": {
            "noauthAuth": {
                "type": "http",
                "scheme": "noauth"
            },
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ],
...
    "paths": {
        "/api/v1/register/": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "create user",
                "description": "### sign up user:\n* Provide user_type if you want to make it admin\n{\n    \"username\": \"admin\",\n    \"password\": \"admin\",\n    \"user_type\": \"admin\"\n}",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "username": "admin2",
                                    "password": "admin"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "CREATED",
                        "headers": {
                            "Server": {
                                "schema": {
                                    "type": "string",
                                    "example": "gunicorn"
                                }
                            },
                            "Date": {
                                "schema": {
                                    "type": "string",
                                    "example": "Sun, 28 May 2023 04:46:17 GMT"
                                }
                            },
                            "Connection": {
                                "schema": {
                                    "type": "string",
                                    "example": "close"
                                }
                            },
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Content-Length": {
                                "schema": {
                                    "type": "integer",
                                    "example": "44"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "User registered successfully."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login/": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "login user",
                "description": "## login user\nProvide username and password for it \n{\n    \"username\": \"admin\",\n    \"password\": \"admin\"\n}",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "username": "admin2",
                                    "password": "admin"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "noauthAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                            "Server": {
                                "schema": {
                                    "type": "string",
                                    "example": "gunicorn"
                                }
                            },
                            "Date": {
                                "schema": {
                                    "type": "string",
                                    "example": "Sun, 28 May 2023 04:47:20 GMT"
                                }
                            },
                            "Connection": {
                                "schema": {
                                    "type": "string",
                                    "example": "close"
                                }
                            },
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Content-Length": {
                                "schema": {
                                    "type": "integer",
                                    "example": "144"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwdWJsaWNfaWQiOjEsImV4cCI6MTY4NTI1MTA0MH0.60FEAFM4lowbNrwXvvFBAk7On3UqWY0Nsj-Hp9u1_nc"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/categories/": {
            "post": {
                "tags": [
                    "Category"
                ],
                "summary": "Create category",
                "description": "#### Crete category for Questions",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "name": "this is cate 5",
                                    "description": "this is cate 5"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwdWJsaWNfaWQiOjEsImV4cCI6MTY4NTI4MDIwNH0.qB4vu47kp0dIXGhkXtjhI0Sq_5R8K-tOyJ9KHW3lDNc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                            "Server": {
                                "schema": {
                                    "type": "string",
                                    "example": "gunicorn"
                                }
                            },
                            "Date": {
                                "schema": {
                                    "type": "string",
                                    "example": "Sun, 28 May 2023 04:12:54 GMT"
                                }
                            },
                            "Connection": {
                                "schema": {
                                    "type": "string",
                                    "example": "close"
                                }
                            },
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Content-Length": {
                                "schema": {
                                    "type": "integer",
                                    "example": "83"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": {
                                        "description": "This is description 2",
                                        "id": 2,
                                        "name": "This is cate 2"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Get all category",
                "description": "**Get all category**",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJwdWJsaWNfaWQiOjEsImV4cCI6MTY4NTI4NDEzM30.IkREq6rZb2YlItk7tpJi5lEKV0j2xQhLs_4D62orK-w"
                    }
                ],
               
                "responses": {
                    "200": {
                        "description": "OK",
                        "headers": {
                            "Server": {
                                "schema": {
                                    "type": "string",
                                    "example": "gunicorn"
                                }
                            },
                            "Date": {
                                "schema": {
                                    "type": "string",
                                    "example": "Sun, 28 May 2023 04:12:54 GMT"
                                }
                            },
                            "Connection": {
                                "schema": {
                                    "type": "string",
                                    "example": "close"
                                }
                            },
                            "Content-Type": {
                                "schema": {
                                    "type": "string",
                                    "example": "application/json"
                                }
                            },
                            "Content-Length": {
                                "schema": {
                                    "type": "integer",
                                    "example": "83"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": {
                                        "description": "This is description 2",
                                        "id": 2,
                                        "name": "This is cate 2"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/user-score-summery/": {
            "get": {
                "tags": [
                    "score summery"
                ],
                "summary": "get all summery",
                "description": "Endpoint: Get Summary of User Quiz Scores\n\nDescription: This endpoint retrieves the summary of a user's scores for a specific quiz.\n\nHTTP Method: GET\n\nQuery Parameters:\n\n- user_id (obtained from current user): The ID of the user for whom to retrieve the quiz scores.\n- quiz_id: The ID of the quiz to retrieve the scores for.\n    \n\n- 200 OK: The quiz scores were successfully retrieved.\n- 400 Bad Request: Invalid user ID or missing quiz ID.\n- 401 Unauthorized: User authentication failed.\n- 404 Not Found: User or quiz not found.\n    \n\nResponse Example (Successful):\n\n```\nHTTP/1.1 200 OK\nContent-Type: application/json\n{\n  \"data\": 25,\n  \"user\": \"john_doe\"\n}\n\n```\n\nResponse Example (No Scores):\n\n```\nHTTP/1.1 200 OK\nContent-Type: application/json\n{\n  \"data\": [],\n  \"user\": \"john_doe\"\n}\n\n```\n\nResponse Example (Action Not Permitted):\n\n```\nHTTP/1.1 400 Bad Request\nContent-Type: application/json\n{\n  \"data\": \"You cannot perform this action\"\n}\n\n```",
                "parameters": [
                    {
                        "name": "quiz_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
...
    }
}